requirements.txt
# Core ML/DL
torch>=2.1.0
transformers>=4.40.0
accelerate>=0.24.0
datasets>=2.14.0

# Quantization & Optimization
bitsandbytes>=0.41.0
optimum>=1.13.0
onnx>=1.14.0
onnxruntime>=1.16.0
autoawq>=0.2.0  # For AWQ quantization

# Fine-tuning
peft>=0.10.0
trl>=0.7.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Data Collection
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0  # For HTML parsing

# Utilities
tqdm>=4.66.0
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Logging & Monitoring
loguru>=0.7.0
faiss-cpu\nbert-score
